test: fix snapshot test of macro expansion to a specific OS
We have to specify a specific OS target because until PyO3 v0.22, PyO3
transitively depends on the [rust-ctor](https://crates.io/crates/ctor) crate,
which generates different output on different OSes.  Once we're doing *that*, we
have to specify a specific Python ABI so that PyO3 doesn't get alarmed about
cross-compilation.  This is all a minor headache.

One concern about this is that everyone testing locally has to run `rustup
target add x864_64-unknown-linuxgnu`.  CI doesn't, because I deliberately picked
the triple that was running on CI, but any local development that isn't on that
OS does.  Is this worth it?